Factor out test entities into separate files
