+++ client-runtime/src/test/java/com/microsoft/rest/HttpBinHeaders.java
@@ -9,6 +9,7 @@
 package com.microsoft.rest;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.microsoft.rest.DateTimeRfc1123;
 
 /**
  * Defines headers for httpbin.org operations.
