+++ client-runtime/src/test/java/com/microsoft/rest/entities/SignedIdentifiersWrapper.java
@@ -1,0 +1,26 @@
+package com.microsoft.rest.entities;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
+import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement;
+
+import java.util.List;
+
+@JacksonXmlRootElement(localName = "SignedIdentifiers")
+public class SignedIdentifiersWrapper {
+    @JacksonXmlProperty(localName = "SignedIdentifier")
+    private final List<SignedIdentifierInner> signedIdentifiers;
+    @JsonCreator
+    public SignedIdentifiersWrapper(@JsonProperty("signedIdentifiers") List<SignedIdentifierInner> signedIdentifiers) {
+        this.signedIdentifiers = signedIdentifiers;
+    }
+    /**
+     * Get the SignedIdentifiers value.
+     *
+     * @return the SignedIdentifiers value
+     */
+    public List<SignedIdentifierInner> signedIdentifiers() {
+        return signedIdentifiers;
+    }
+}
