+++ client-runtime/src/test/java/com/microsoft/rest/entities/Slide.java
@@ -1,0 +1,15 @@
+package com.microsoft.rest.entities;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
+
+public class Slide {
+    @JacksonXmlProperty(localName = "type", isAttribute = true)
+    public String type;
+
+    @JsonProperty("title")
+    public String title;
+
+    @JsonProperty("item")
+    public String[] items;
+}
