+++ client-runtime/src/test/java/com/microsoft/rest/http/MockHttpClient.java
@@ -8,7 +8,6 @@ package com.microsoft.rest.http;
 
 import com.google.common.io.CharStreams;
 import com.microsoft.rest.Base64Url;
-import com.microsoft.rest.Base64UrlTests;
 import com.microsoft.rest.DateTimeRfc1123;
 import com.microsoft.rest.HttpBinJSON;
 import com.microsoft.rest.policy.RequestPolicy;
